Extract all non-specific parts to simplexmq
