+++ cabal.project
@@ -9,7 +9,7 @@ constraints: zip +disable-bzip2 +disable-zstd
 source-repository-package
     type: git
     location: https://github.com/simplex-chat/simplexmq.git
-    tag: 0410948b56ea630dfa86441bbcf8ec97aeb1df01
+    tag: e4ae2a5faf5ba3042980402c934119e9c26742f9
 
 source-repository-package
     type: git
