+++ package.yaml
@@ -36,7 +36,6 @@ dependencies:
   - mtl == 2.3.*
   - network >= 3.1.2.7 && < 3.2
   - network-transport == 0.5.6
-  - network-udp >= 0.0 && < 0.1
   - optparse-applicative >= 0.15 && < 0.17
   - process == 1.6.*
   - random >= 1.1 && < 1.3
