+++ simplex-chat.cabal
@@ -1,6 +1,6 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.35.2.
+-- This file has been generated from package.yaml by hpack version 0.36.0.
 --
 -- see: https://github.com/sol/hpack
 
@@ -129,11 +129,6 @@ library
       Simplex.Chat.ProfileGenerator
       Simplex.Chat.Protocol
       Simplex.Chat.Remote
-      Simplex.Chat.Remote.Discovery
-      Simplex.Chat.Remote.Multicast
-      Simplex.Chat.Remote.Protocol
-      Simplex.Chat.Remote.Transport
-      Simplex.Chat.Remote.Types
       Simplex.Chat.Store
       Simplex.Chat.Store.Connections
       Simplex.Chat.Store.Direct
@@ -151,6 +146,7 @@ library
       Simplex.Chat.Terminal.Output
       Simplex.Chat.Types
       Simplex.Chat.Types.Preferences
+      Simplex.Chat.Types.Remote
       Simplex.Chat.Types.Util
       Simplex.Chat.Util
       Simplex.Chat.View
@@ -183,7 +179,6 @@ library
     , mtl ==2.3.*
     , network >=3.1.2.7 && <3.2
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
@@ -236,7 +231,6 @@ executable simplex-bot
     , mtl ==2.3.*
     , network >=3.1.2.7 && <3.2
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
@@ -290,7 +284,6 @@ executable simplex-bot-advanced
     , mtl ==2.3.*
     , network >=3.1.2.7 && <3.2
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
@@ -346,7 +339,6 @@ executable simplex-broadcast-bot
     , mtl ==2.3.*
     , network >=3.1.2.7 && <3.2
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
@@ -401,7 +393,6 @@ executable simplex-chat
     , mtl ==2.3.*
     , network ==3.1.*
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
@@ -460,7 +451,6 @@ executable simplex-directory-service
     , mtl ==2.3.*
     , network >=3.1.2.7 && <3.2
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
@@ -545,7 +535,6 @@ test-suite simplex-chat-test
     , mtl ==2.3.*
     , network ==3.1.*
     , network-transport ==0.5.6
-    , network-udp ==0.0.*
     , optparse-applicative >=0.15 && <0.17
     , process ==1.6.*
     , random >=1.1 && <1.3
