+++ src/Simplex/Chat/Store/Shared.hs
@@ -29,8 +29,8 @@ import qualified Database.SQLite.Simple as SQL
 import Database.SQLite.Simple.QQ (sql)
 import Simplex.Chat.Messages
 import Simplex.Chat.Protocol
-import Simplex.Chat.Remote.Types
 import Simplex.Chat.Types
+import Simplex.Chat.Types.Remote
 import Simplex.Chat.Types.Preferences
 import Simplex.Messaging.Agent.Protocol (AgentMsgId, ConnId, UserId)
 import Simplex.Messaging.Agent.Store.SQLite (firstRow, maybeFirstRow)
