+++ tests/JSONTests.hs
@@ -9,7 +9,7 @@ import qualified Data.ByteString.Lazy.Char8 as LB
 import GHC.Generics (Generic)
 import Generic.Random (genericArbitraryU)
 import MobileTests
-import Simplex.Chat.Remote.Protocol (owsf2tagged)
+import Simplex.RemoteControl.Protocol (owsf2tagged)
 import Simplex.Messaging.Parsers
 import Test.Hspec
 import Test.Hspec.QuickCheck (modifyMaxSuccess)
